Add Delete Functionality

+++ DCAUA/public/js/edit_form_delay.js
@@ -9,7 +9,20 @@ $(document).ready(function () {
         usedmethod.submitEditForm($('#submit_edit_form'), '/delay_compensation_form_list/update_edit_form', $(this));
     });
     $(document).on('click', '#delete_form_btn', async function (e) {
-        usedmethod.deleteForm($(this), '/delay_compensation_form_list/delete_form');
+        e.preventDefault();
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "Do You Want To Submit It",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, Submit it!'
+        }).then(async (result) => {
+            if (result.isConfirmed) {
+                usedmethod.deleteForm($(this), '/delay_compensation_form_list/delete_form');
+            }
+        });
     });
 });
